update: Clean up misuse of wu() in prod_nat()
Don't use multiple calls of wu() to print a single line, because that
creates a separate telegram for each part.  The read command normally
merges the telegrams, but if they are more than five seconds apart
(clock jumped somehow), we get a telegram header in the middle of a
line.  Unlikely to happen, but clean it up anyway.

The misuse has always been there.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>