news: Merge news more aggressively
News reporting merges news items into recent items with same actor,
action and victim, as long as the merged number of times doesn't
exceed 65535 (127 before the previous commit).

This complicates incremental xdump a bit: when the client sees a row,
it's not obvious which of the previously dumped rows it replaces.
Actually, rows with the same actor, action, victim and time replace in
xdump order.

Merge news item regardless of the number of times; saturate the number
at 65535.  Now the replacement is obvious, because there can only be
one news item with the same actor, action, victim and time.

This also rate-limits floods of identical news should they happen for
some reason.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>