attack assault paradrop: Execute spies on takeover
When the attacker takes a sector, spies try to hide.  If they fail,
they're treated like any other land unit: they attempt to
self-destruct, and if the damage isn't lethal, they get captured.
Successful self-destruct is reported as "blown up by the crew", which
makes no sense for spies.  Spies surviving self-destruct is odd, as
any damage is normally fatal for them.

Extend the special case for spies: summarily execute the ones that
fail to hide.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>