info/Wolfpack: Redundant, delete
The .TH and .NA header promise information about "Wolfpack Code" and
"The Wolfpack Project", but the body doesn't really deliver.  It's
basically the first paragraph of Empire4.t plus pointers to Options.t
and Empire.t.  Has been that way since it was added in 4.0.2.
History.t covers the Wolfpack project more usefully, so delete this
one.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>