travis: Specify dist to fix Linux build
The build fails on Travis since we require autoconf 2.69 (commit
db055a).  Travis still defaults to Ubuntu Precise, and Precise's
autoconf is too old.  We could revert commit db055a and rely on Travis
to keep things working with old autoconf.  But Precise reached end of
life a couple of months ago, and I suspect it's only a matter of time
until Travis drops it, too.

Fix the build on Travis by requesting Trusty.  Its end of life is
planned for April 2019.

Signed-off-by: Markus Armbruster <armbru@pond.sub.org>